fix: removed uconfig dependency (#17)
The package uconfig has a problem with a retracted latest version. This
was causing some problem when running `go get -u ./...`.

We only using a single function from that package so we copied the
function to this repository and removed the external dependency. This
was copied from the `go-app` project.

An example was added based on the official Minimal Word Count to show
how a pipeline can be integrated.

Some dependencies were updated and some log messages fixed.